LayoutTests/accessibility/aria-readonly-expected.txt

This tests that the readonly state of the AXValue property is correctly reported for native and non-native elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Elements to test: 47

PASS htmlEditableDivIsWritable is true
PASS htmlEditableDiv2IsWritable is false
PASS htmlEditableDiv3IsWritable is true
PASS htmlEditableDiv4IsWritable is true
PASS htmlEditableDiv5IsWritable is false
PASS htmlEditableDiv6IsWritable is true
PASS htmlNonEditableDivIsWritable is false
PASS htmlNonEditableDiv2IsWritable is false
PASS htmlNonEditableDiv3IsWritable is true
PASS ariaTextBoxIsWritable is true
PASS ariaReadOnlyAriaTextBoxIsWritable is false
PASS htmlReadOnlyTextFieldIsWritable is false
PASS htmlReadOnlyTextField2IsWritable is false
PASS htmlReadOnlyTextField3IsWritable is false
PASS htmlReadOnlyTextField4IsWritable is false
PASS htmlReadOnlyTextAreaIsWritable is false
PASS htmlReadOnlyTextArea2IsWritable is false
PASS htmlReadOnlyTextArea3IsWritable is false
PASS htmlReadOnlyTextArea3IsWritable is false
PASS textFieldIsWritable is true
PASS ariaReadOnlyTextFieldIsWritable is true
PASS ariaNonReadOnlyTextFieldIsWritable is true
PASS textAreaIsWritable is true
PASS textArea2IsWritable is true
PASS textArea3IsWritable is true
PASS ariaGridCellIsWritable is false
PASS ariaGridCell2IsWritable is true
PASS ariaColumnHeaderIsWritable is false
PASS ariaColumnHeader2IsWritable is true
PASS ariaRowHeaderIsWritable is false
PASS ariaRowHeader2IsWritable is true
PASS ariaGridIsWritable is false
PASS ariaGrid2IsWritable is true
PASS ariaTreeGridIsWritable is false
PASS ariaTreeGrid2IsWritable is true
PASS ariaGridCell3IsWritable is false
PASS ariaGridCell4IsWritable is true
PASS ariaGridCell5IsWritable is false
PASS ariaGridCell6IsWritable is true
PASS ariaColumnHeader3IsWritable is false
PASS ariaColumnHeader4IsWritable is true
PASS ariaColumnHeader5IsWritable is false
PASS ariaColumnHeader6IsWritable is true
PASS ariaRowHeader3IsWritable is false
PASS ariaRowHeader4IsWritable is true
PASS ariaRowHeader5IsWritable is false
PASS ariaRowHeader6IsWritable is true
PASS successfullyParsed is true

TEST COMPLETE
